Add read, update and delete query methods, add sql file, update index of gateway to export gateway, update server routes to no longer console log and sends correct status, add appropriate tests for postgres plugins and gateway.

+++ src/gateways/schema/schema.sql
@@ -3,7 +3,7 @@ CREATE DATABASE inventory
 CREATE TABLE IF NOT EXISTS products (
   product_id varchar(250) NOT NULL,
   style_id varchar(250) NOT NULL,
-  size integer NOT NULL,
+  size numeric NOT NULL,
   quantity integer NOT NULL
   PRIMARY KEY (product_id)
 )
